centralizing schema management into single file service
